+++ app/src/main/java/labs/zero_one/roundo/MapKit.kt
@@ -5,6 +5,7 @@ import android.content.res.Resources
 import android.location.Location
 import android.os.Bundle
 import com.minemap.minemapsdk.MinemapAccountManager
+import com.minemap.minemapsdk.annotations.Marker
 import com.minemap.minemapsdk.annotations.MarkerOptions
 import com.minemap.minemapsdk.camera.CameraPosition
 import com.minemap.minemapsdk.camera.CameraUpdateFactory
@@ -18,11 +19,16 @@ import com.minemap.minemapsdk.maps.MineMap
  * ## 属性列表
  * - [mineMap]
  * - [isMapInitialized]
+ * - [markerList]
+ * - [tempMarkerOptionsList]
  * - [isFollowing]
  *
  * ## 方法列表
  * - [initMap]
  * - [moveTo]
+ * - [add]
+ * - [addMarkerAt]
+ * - [clearMarkers]
  *
  * @param [context] 环境
  *
@@ -31,12 +37,16 @@ import com.minemap.minemapsdk.maps.MineMap
  *
  * @property [mineMap] 地图控制器
  * @property [isMapInitialized] [mineMap] 是否成功初始化
+ * @property [tempMarkerOptionsList] 标记暂存列表
+ * @property [markerList] 标记列表
  * @property [isFollowing] 中心是否跟随移动
  */
 class MapKit(private val context: Context) {
 
     private lateinit var mineMap: MineMap
     private var isMapInitialized = false
+    private var markerList: ArrayList<Marker> = ArrayList(0)
+    private var tempMarkerOptionsList: ArrayList<MarkerOptions> = ArrayList(0)
     var isFollowing = true
 
     /**
@@ -58,8 +68,12 @@ class MapKit(private val context: Context) {
         mapView.onCreate(savedInstanceState)
         mapView.getMapAsync { newMap ->
             if (newMap == null) throw Exception(context.getString(R.string.err_map_init_failed))
-            mineMap = newMap
             isMapInitialized = true
+            mineMap = newMap
+            for (markerOptions in tempMarkerOptionsList) {
+                markerList.add(mineMap.addMarker(markerOptions))
+            }
+            tempMarkerOptionsList.clear()
             mineMap.setStyleUrl("http://minedata.cn/service/solu/style/id/4810")
             mineMap.uiSettings.isCompassEnabled = true
             mineMap.uiSettings.setCompassMargins(
@@ -72,9 +86,8 @@ class MapKit(private val context: Context) {
             mineMap.cameraPosition =
                 CameraPosition.Builder()
                     .target(LatLng(locationKit.lastLocation))
-                    .zoom(16.0)
+                    .zoom(12.0)
                     .build()
-
         }
     }
 
@@ -99,14 +112,21 @@ class MapKit(private val context: Context) {
     /**
      * 添加标记
      *
-     * @param [marker] 标记
+     * ## Changelog
+     * ### 0.1.9
+     * - 若地图还未初始化则将标记存入 [tempMarkerOptionsList]，待初始化后再加入
+     *
+     * @param [markerOptions] 标记
      *
      * @author lucka-me
      * @since 0.1.7
      */
-    fun add(marker: MarkerOptions) {
-        if (!isMapInitialized) return
-        mineMap.addMarker(marker)
+    fun add(markerOptions: MarkerOptions) {
+        if (isMapInitialized) {
+            markerList.add(mineMap.addMarker(markerOptions))
+        } else {
+            tempMarkerOptionsList.add(markerOptions)
+        }
     }
 
     /**
@@ -117,7 +137,20 @@ class MapKit(private val context: Context) {
      * @author lucka-me
      * @since 0.1.7
      */
-    fun addMarkAt(location: Location) {
+    fun addMarkerAt(location: Location) {
         add(MarkerOptions().position(LatLng(location)))
+    }
+
+    /**
+     * 清空标记
+     *
+     * @author lucka-me
+     * @since 0.1.9
+     */
+    fun clearMarkers() {
+        for (marker in markerList) {
+            mineMap.removeMarker(marker)
+        }
+        markerList.clear()
     }
 }